Enjoy the output of math performed with list items
Given an integer n, create a list of integers from 2 to n inclusive. Print the multiples of every integer of the list, starting from 2 up to half the length of the list. Print the remaining half of the integers as it is and ignore the last integer. The multiples and individual integers must be less than or equal to the value of n and must appear between the statements "Block start" and "Block end."