Upgrade or retain the scores with a little math
This piece of code is fun. The program accepts two integer inputs, the first line of input accepts the length of the array and the following lines are the grades in the array. The array is passed as an argument to a function where the scores are checked against few rules.

If the difference between the grade and the next multiple of five is less than three, then the grade is rounded to the next multiple of five. If the grade is less than 38 or the difference between grade and the next multiple of five is greater than three the grades are retained. Try it yourself!