2D Matrix Difference of Diagonal Sum
Concepts involved are list, math operators and operations
The program accepts two inputs. The first line accepts the size of the matrix and the following lines will accept the matrix elements. 
The for loop within the function adds the diagonal elements into left and right list. For a matrix of odd lengths, the overlapping element is removed and the sum difference of the elements is returned.  Try it yourself!