Get Diagonal, skip first row and column
This program is an accidental discovery. But it is fun, you must try too. I was trying to get the sum of diagonals and find their difference. It turned out that I got the left diagonal elements and then I got all the elements of matrix skipping first row and columns. 
The program accepts two inputs, the first line accepts the length of the matrix and the following lines accept the matrix elements. the nested for loop makes the difference. Try it yourself!