Partial solution for magic matrix
We all know the magic matrix has a pattern for integer placement within the matrix. Integers in the center row and center column should all be an odd number, the left top, right top, left bottom and right bottom integers should be even numbers in a 3 X 3 matrix. My approach to design a solution for the Magic Matrix problem was to rewrite the given matrix by replacing any integer with zero if they are violating the integer placement rule and if the integers are repeating. The above piece of code is based on the facts that the integer at [1][1] should be five, the sum of the matrix has to be 45 and it should have integers ranging from 1-9 inclusive. My idea was to create a list for missing integers from the rewritten matrix, replace all the zeros with the from the list of missing integers. This solution works partially but when it comes to swapping the integer placement there is no ideal solution. At least not now.