Split the array and find the number of same items
Given two arrays, the first array consists of n integers, and the second array consists of a sequence of integers, referenced as the index of the first array. For the sake of this challenge, the index of the first array assumed to be starting with 1. The task is to find the number of identical integers in the first array considering the item value in the second array as the index reference for the first array. 
I have used the slicing approach to solve this problem.  Try it yourself!