Why choose list comprehension over nested loops
The task was to implement list comprehension for printing the coordinates of a cuboid. The constraints were, given 4 integers, the sum of the first three integers should not be the same as 4th integer. Because of for loop the upper limit of generated integers is taken care of. Try it yourself!