Perform list operations via input commands
The task was to perform list operations. The program accepts two inputs; the first input N is the number of activities, and the second input is the sequence of commands to execute on the list. With every command, I will look for keywords that match with list operation commands; accordingly, I will perform the activities.