rjust() ljust() center()
Usually, the codes that I write my logic go on to my git repo. This code is an exception. I just had to experiment with rjust() ljust() or center() to print a HackerRank text logo. So, this is the solution if you want to save some time in completing this challenge.