Change to upper case on two conditions
The task was to convert lowercase characters to uppercase. The given string can be alphanumeric and will contain white space. We know the string is immutable, one way to solve this problem is by using the list method. However, it will increase the computational time. The first character in the string, by default, has to be an upper case, change character at position i = 0 to uppercase. Then, for every space found at position i the character at i+1 should be an upper case. For the rest of the conditions, continue to concatenate a string to a new empty string.