Count unique string in an array, two methods.
The task was to find the count of unique strings in a given array. There is more than one way of solving this problem. In the first method, I used the Brute force. For every element supplied and before it gets appended,  the function will compare if there are such strings that exist before and get the count. If yes, the string gets ignored. The second method is self-explanatory. Numpy does the job.