Pile the cubes as Cube-j length >= Cube-i
I would call this my celebrated code for this week. I was happy when I saw this code passed all the test cases.
The task was to pile the cubes. When stacking the cubes, you can only pick up either the leftmost or the rightmost cube each time. Print "Yes" if it is possible to stack the cubes. Otherwise, print "No". Try this yourself!