List to dictionary, sort by value with key ref
The task is to find the characters with the highest frequency in a given list. If the frequency is similar for two or more strings, sort the characters alphabetically. The program accepts the string input, creates a  list out of the string provided. With the created list, the set function will retain the unique characters and then sorted will sort the order alphabetically. Next, a dictionary is created with character as key and frequency as value. The second for loop will loop through the sorted dictionary. This will help print the content alphabetically.