Max of array multiplied by its length
The task was to map alphabets given the word with its integer value provided as input, later find the max value associated with the alphabet of the word, and multiply it with the word length. I created a dictionary, using a for loop map the integer value with the alphabets and use max function to get the max value. The result will be max(height) X len(word)

Try it yourself!