I thought Binary insertion sort was the solution
7/11 test cases passed
I might change the solution to this challenge. The above code is a partial solution. I tried implementing Binary Insertion sort to get rid of Terminated due to timeout :( but no luck. So far, I have not come across a 100% solution. But this is my code. Check my Medium article to know the working of the code.