HackerRank, Equilize the Array
The task is to find the difference between the length of the given array, and the length of the element, that has max repetition. For example; arr = [a,b,c,c,d,c,c,s] the length of the given array is 8 the. Character 'c' repeats 4 times which is the maximum number of repetitions. 8-4 = 4 We need to remove the 4 other elements from the given array to equalize the array.