HackerRank, Viral Advertising
The task was to get the right sum of an array. How we calculate the sum is the tricky part. 

On the first day, half of those 5 people (i.e., floor(5/2) = 2 ) like the advertisement and each shares it with 3 of their friends. At the beginning of the second day,  floor(5/2) x 3 = 6 people receive the advertisement.