Create Repeating_Str.py
Lilah has a string s, of lowercase English letters that she repeated infinitely many times. Given an integer n, find and print the number of letter a's in the first

letters of Lilah's infinite string.

For example, if the string s = 'abcac'
and n = 10, the substring we consider is abcacabcac, the first characters of her infinite string. There are 4 occurrences of a in the substring.

Function Description

Complete the repeatedString function in the editor below. It should return an integer representing the number of occurrences of a in the prefix of length
in the infinitely repeating string